requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
langchain==0.1.0
langchain-openai==0.0.5
qdrant-client==1.7.0
python-multipart==0.0.9
pydantic==2.6.1
python-docx==1.0.1
PyPDF2==3.0.1
pytest==7.4.4
httpx==0.26.0  # For async HTTP requests
python-jose==3.3.0  # For encryption
loguru==0.7.2   # For better logging